saveEmployerWithoutImage and unit test

+++ src/main/java/com/lifepill/employeeService/dto/EmployerWithoutImageDTO.java
@@ -16,7 +16,7 @@ import java.util.Date;
 @Data
 public class EmployerWithoutImageDTO {
     private long employerId;
-    private long branchId;
+    private int branchId;
     private String employerNicName;
     private String employerFirstName;
     private String employerLastName;
